Port PageView to spina.
This change ports the base PageView class to spina and TypeScript. As
part of this, the code has been moved into the `common` subdirectory to
match the bundle that it's a part of.

The one change that was required outside of this class was to update the
`DiffViewerPageModel` to use `preinitialize` instead of `constructor`.
The use of the reserved name `constructor` has bitten us in the past,
and wasn't working anymore now that the base class was an ES6 class.

Testing Done:
- Ran js-tests.
- Verified that page views still rendered correctly.

Reviewed at https://reviews.reviewboard.org/r/12937/