Make it possible to import jasmine.
Test files using jasmine were extremely ugly with lots of errors,
because all of the jasmine methods were just loaded as globals. This
change adds the relevant rollup configuration so that we can import
those methods and properly have them get mapped to the items on
`window`.

This also fixes the package dependency to be `jasmine-core`, which is
the correct name on npm. The `jasmine` npm package is a CLI tool for
running tests, rather than the library.

Testing Done:
Used this in conjunction with changes that did
`import { ... } from 'jasmine'` and saw that all unit tests ran
correctly.

Reviewed at https://reviews.reviewboard.org/r/12935/