Fix API issues creating comments on draft diffs.
This change fixes up a few issues with the API making comments on
not-yet-published diffs. The main problem here is that the DiffSet is not
added to the review request's DiffSetHistory until it is published. This
change fixes up the queries for both `get_queryset` and `create` to
handle this.

Testing Done:
- Ran unit tests.
- Added some comments to a diff that was not yet published.

Reviewed at https://reviews.reviewboard.org/r/12946/