Make the unified banner show the draft + review state.
The current mode calculation for the unified banner assumed that if the
review request was not yet public, that that was the only possible
state. This change fixes it up so we can also have a pending review.

Testing Done:
Created a review on an unpublished review request and saw the unified
banner show the correct states in the mode selector.

Reviewed at https://reviews.reviewboard.org/r/12944/