Prevent visual jumps when editing the change description.
The line-height of the label for the change description field in the
unified banner wasn't enforcing a match between the text and the edit
icon, which meant that the UI would jump a few pixels when entering edit
mode (which hides the icon). Setting this explicitly makes it so things
don't reflow when that icon appears or disappears.

Testing Done:
Entered/exited edit mode for the change description and saw that
everything stayed in one place.

Reviewed at https://reviews.reviewboard.org/r/12953/