Plumb through redirect URL on the "Log in with SAML SSO" button.
When SAML autologin is enabled, we were correctly passing through the
redirect URL, but that was not included on the link for the "Log in with
SAML SSO" button on the login page. This was annoying to users when they
tried to open links to specific pages and they'd end up on the dashboard
instead.

Testing Done:
- Ran unit tests.
- Opened /r/1/ on my test server and saw that I was redirected to the
  login page. Clicked the SAML login button, went through that flow, and
  saw that I was correctly redirected back to /r/1/ instead of to
  /dashboard/

Reviewed at https://reviews.reviewboard.org/r/12957/