Fix API issues creating comments on draft file attachments.
This change fixes up a couple issues with the API making comments on
not-yet-published file attachments. The main problem here is that draft
attachments are related only to the draft and not yet to the review
request. This change fixes up the questios for both `get_queryset` and
`create` to handle this.

Testing Done:
- Ran unit tests.
- Added some comments to file attachments which were not yet published.

Reviewed at https://reviews.reviewboard.org/r/12962/