Improve related object selectors in conditions.
The object selectors in the conditions UI leave a fair bit to be
desired. They end up rendering as `<select multiple>`, which isn't a
particularly great UI, and the results were displaying in whatever order
the database was giving us.

I'm continuing to investigate whether we have better options for the
widgets, but for now this change fixes up the ordering for repositories,
repository types, and groups to sort alphabetically instead of randomly.

Testing Done:
Verified that related object selectors in conditions had better
ordering.

Reviewed at https://reviews.reviewboard.org/r/12987/