Fix bugs in SSO autologin.
The port of the SSO autologin to 5.0.x had a couple problems. First, we
were using a bad reference for the `ItemLookupError` check. Second,
django had deprecated `is_safe_url` and inadventantly changed the way it
works in the process. This change updates the code to use its successor
in the correct way.

Testing Done:
Enabled SSO auto-login and was able to successfully go through the login
flow.

Reviewed at https://reviews.reviewboard.org/r/12999/