Make entire items in the review menu clickable.
The items in the review menu were only clickable within the bounding box
of their text. This change fixes up the styles so that the `<a>`
elements occupy the full size of the menu item.

Testing Done:
Verified that clicking in empty space where an item was still
highlighted triggered the action.

Reviewed at https://reviews.reviewboard.org/r/13009/