Save open review reply editors before publishing.
This change adds review reply editors to the set of things that will be
saved when publishing new content from the unified banner. Incidentally,
it appears that the old draft reply banners never did this (meaning we'd
fail to save new reply content when hitting the publish button).

This change also fixes up a couple regressions in event names for legacy
actions which are hidden when the unified banner feature is enabled.

Testing Done:
- Ran js-tests.
- Set some pending data in replies to review headers and comments and
  then published from the unified banner. Saw the open editors save
  correctly and the reloaded page contain the new content.

Reviewed at https://reviews.reviewboard.org/r/13044/