Fix minimum height of comment reply editors.
In the update to the review reply editors to switch to
`InlineEditorView`, I read the existing code a little bit wrong, and
mistakenly overrode the minimum height default to be zero. This change
removes that override, so that the default height for reply editors is
the default editor minimum (70px).

Testing Done:
Opened editors for replies to comments and review headers and saw the
correct minimum editor height.

Reviewed at https://reviews.reviewboard.org/r/13044/