Don't reload the page when merely closing the review dialog.
We had recently added a change to explicitly reload the page after
publishing reviews from the review dialog. This was done because with
the new unified banner, we no longer had the post-publish trigger that
was causing the reload. Unfortunately, this was also causing the page to
reload when the "Close" button was clicked. This change fixes it so we
only do it when publishing.

Testing Done:
Opened the review dialog, edited some comments, and then clicked
"Close". Saw the review save and dialog close without the page
reloading.

Reviewed at https://reviews.reviewboard.org/r/13055/