Add typing for reviewboard.hostingsvcs.repository.
This introduces typing for `RemoteRepository`, which represents a
scanned repository in a service. This covers typing of the attributes on
the class and on the `__init__()` and `__repr__()` methods.

The module has also been given a docstring.

Testing Done:
Viewed the docs in the new docs UI.

Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/12989/