Fix style regression for diff change descriptions.
The change descriptions for diffs use the same `.diff-index` class as
the file listing in the diffs. Unfortunately, some of the styles for
that class were moved to the listing's parent container in order to
facilitate the new docking behavior, which caused the look of diff
change descriptions to regress.

This change copies over the few styles we need in order for this to look
correct.

Testing Done:
Published some updated diffs and saw the "Diff" entry in the change
description box looked correct.

Reviewed at https://reviews.reviewboard.org/r/13057/