Add a JavaScript hook for extending the File Attachment thumbnail container UI.
This adds a JavaScript hook for extending the File Attachment thumbnail
container UI. The hook can be used to display additional content on
file attachment thumbnails and in the file attachment actions menu.

This also adds a method to the File Attachment thumbnail view that allows
callers to add an action to the file attachment actions menu.

Testing Done:
- Ran JS unit tests.
- Ran unit tests.
- Manually tested with the new "Download as PDF" action that's being added in
  in Power Pack.
- Manually tested with a simple extension and view, the one in the example
  in the docs.
- Built the docs and made sure they look ok.

Reviewed at https://reviews.reviewboard.org/r/12951/