Create pytest infrastructure in `rbext create`
Now that we can use pytest in extensions, we want to create that as part
of creating a new extension. This change adds the machinery to generate
`conftest.py` and `pyproject.toml` files with the correct setup to run
tests.

Testing Done:
Created a test extension and saw that it was set up to run tests.

Reviewed at https://reviews.reviewboard.org/r/13084/