Fix site version parsing in rb-site upgrade with modern packaging libs.
When running `rb-site upgrade`, the old site version is checked to see
if certain steps need to be performed. We check the display version of
the site, since that's all that we can pull from siteconfig, and
historically this has worked, with `pkg_resources.parse_version()`
turning this into a `LegacyVersion` object.

Unfortunately, this support has been dropped, causing this code path to
fail if running any beta, alpha, RC, or development releases.

To work around this, we now pull out the actual version component (the
part before a space) and compare using that.

This is a backport of the relevant parts of the code from commit
766b03221 (minus the removal of `pkg_resources`).

Testing Done:
Tested a site upgrade, verifying that this code path was being executed.
Verified that the version check worked fine on an alpha dev build.

Reviewed at https://reviews.reviewboard.org/r/13086/