Rewrite the FAQ for Review Board.
The old FAQ was very out-of-date, a bit developer-centric, and didn't
cover much of the interesting offerings we have today.

This is a first pass at rewriting the FAQ to be more useful to people
stopping by the site. This is not an exhaustive FAQ, and there's more we
could add to it (going over more Power Pack features, for instance), but
this is only meant to be a starting point for new content.

Testing Done:
Built the docs. Checked for build errors, spelling errors, and bad links.
Viewed it in the new docs UI.

Reviewed at https://reviews.reviewboard.org/r/13006/