Add context parameter to actions JS data methods.
It turns out that it will be useful to allow the `get_js_*_data` for
actions to have access to the rendering context. This change does that
by introducing some new template tags that call the methods with the
context.

This also allows improving the implementation of the menu view a bit,
removing non-rendering actions from the data sent to the JS model,
preventing some warnings about unknown IDs from popping up in the case
that an extension has hidden a menu-item action.

Testing Done:
- Implemented a custom action that used the context parameter.
- Used `HideActionHook` to hide a menu item action and saw that I was no
  longer getting warnings in the JS console.
- Ran js-tests.

Reviewed at https://reviews.reviewboard.org/r/13107/