Port TextEditorView and DnDUploader to spina.
In preparation for integrating the rich text toolbar, this change ports
our `TextEditorView` and `DnDUploader` to TypeScript via spina.

This also fixes one issue I discovered where if the only thing you've
done with a text editor is upload an image, it won't think the editor is
dirty and so saving discards the contents.

Testing Done:
- Ran js-tests.
- Used drag and drop to upload files into a text editor.

Reviewed at https://reviews.reviewboard.org/r/13115/