Fix some inconsistent build issues and legacy configuration for docs.
Every so often, we have some sporadic build issues that don't appear to
make much sense. Mostly it seems to be logging-related, with logged
output causing some blocked I/O issues. A possible culprit is the way we
initialize Review Board in our Sphinx extension.

This change updates that initialization to turn off logging, extensions,
and template management, with a hope that this will avoid any issues
that may be impacting Sphinx.

Doc-related URL settings now use `https` instead of `http` in places,
and some legacy imports have been removed.

Along with that, the release notes now include the manual's extension
for specialized references, so that the release notes can refer to
relevant sections in the manual.

Testing Done:
Successfully built the manual and release notes.

Tested the references with the in-progress changes for 5.0.5 release
notes.

Reviewed at https://reviews.reviewboard.org/r/13116/