Update the crontab generation to produce a better file for virtualenvs.
For virtualenv-based installs, we now just hard-code the path to
rb-site, which will ensure the right Python environment is used.

For a global install, we use the logic we've had before. We could change
this to also point to the rb-site path, but the plan for now is to avoid
altering behavior that's known to work globally.

Testing Done:
Tested the crontabs for both on two fresh sites, making sure they both
ran as expected.

Reviewed at https://reviews.reviewboard.org/r/13118/