Add a formatting toolbar for interacting with markdown fields.
This change adds a markdown formatting toolbar for rich text editors.
The buttons for inline formatting add the corresponding syntax to the
selection if one exists, or just inserts the syntax and positions the
cursor. This applies to the bold, italic, strikethrough, ordered and
unordered list, and code buttons. If the selection begins and ends with
the same syntax as the button being pressed, the syntax is removed. For
example, if the bold button is pressed and the selection is already
bolded, the bold syntax is removed.

For the insert link button, if there is a selection, the selection
becomes the text to display for the url. Otherwise, it inserts syntax
for a link and positions the cursor where the text to display will go.
Positioning the cursor in a formatted link, selecting the text in the
link, or the either link will remove the formatting.

For the upload image button, it opens a dialog to upload an image. This
triggers the same code path as when an image is dropped on the editor,
performing the upload and inserting the markdown reference.

For the ordered and unordered list buttons, numbers (1.) or dashes (-)
are appended to or removed from to the beginning of the current line.

Based on work by Hannah Lin.

Testing Done:
- Ran unit tests.
- Manually tested all of the different formatting options, including
  links and image uploads.

Reviewed at https://reviews.reviewboard.org/r/11004/