Add confirmation dialogs when discarding drafts from the unified banner.
One feature that was missing from the new banner that was present in the
old was a confirmation dialog when discarding a draft review.
Additionally, we didn't have any kind of confirmation step when
discarding review request drafts. This change implements both of those
in the new UI.

Testing Done:
Discarded review request drafts and review drafts. Saw the confirmation
dialog, which canceled the operation through either the escape key or
clicking the cancel button, and proceeded when clicking the discard
button.