Parameterize ActionView and MenuActionView.
In order to make it easier to extend the JS-side actions with full
typing, we need to make the views parameterized so that the model and
options can be altered for subclasses.

Testing Done:
Used in conjunction with an extension that had a custom options
interface on an action view subclass.

Reviewed at https://reviews.reviewboard.org/r/13132/