Catch errors when uninitializing ActionHook.
I recently hit a case where an extension that adds a bunch of actions
was hitting errors during devserver initialization and reloads. The
issue seems to be that it was never fully initialized, then the
devserver reloaded, and hitting an error during the extension shutdown
was leaving old action registrations around, which would then fail due
to duplicates. This seems like it's probably a devserver-specific
problem, but it's annoying. This change catches any ItemLookupErrors
that get triggered during ActionHook shutdown.

Testing Done:
Was able to use my extension.

Reviewed at https://reviews.reviewboard.org/r/13131/