Fix min-height of the review request actions bar.
The actions bar for review requests had a minimum height set, but it was
1 pixel too small. Most of the time this doesn't matter because the
actions set the height, but in the case of viewing a review request
that doesn't have a diff anonymously, the actions bar is actually
completely empty. This resulted in the tabs being shifted down one
pixel.

Testing Done:
Opened a private browser window and looked at a review request that
didn't have a diff. Saw that the review request tabs were no longer
misaligned.

Reviewed at https://reviews.reviewboard.org/r/13151/