Make naming of JS model files consistent.
For a long time we've had a mix of filenames for JS model files, with
some of them ending in "Model" and some of them not. This change makes
all of our new TypeScript code consistent, standardizing on the form
that includes "Model" in the filename. As more things are ported, we'll
fix up the remaining ES6/JS files.

Testing Done:
Ran js-tests.

Reviewed at https://reviews.reviewboard.org/r/13156/