Increase spacing and reduce clutter in the dashboard styling.
The dashboard is quite information-dense, with a lot of information
running together.

In an attempt to give that information more breathing room, this change
adds some additional padding to the header, cells, and paginator, taking
spacing from 4px in the header, 5px in the cells, and 0.7em in the
footer to 0.75em across the board. (Technically, the header is now
0.75em horizontal, 0.5em vertical).

This makes it just a bit easier to focus on individual review request
summaries (particularly when they're very long), and better promotes the
most recent review requests.

There's no longer a border between rows, cutting back that bit of visual
excess.

There's also a fix for the styling of the age datetime borders. The
top-most one always had a top border, which looked bad. The top-most one
is now flush against the header.

Testing Done:
Tested the dashboard and other datagrid pages, comparing between the old
and new styles.

Initially, it felt like perhaps too much whitespace, but it quickly felt
far easier to read and digest the information.

Reviewed at https://reviews.reviewboard.org/r/13150/