Add the tips carousel to the review dialog.
With the change to pop up the review dialog as soon as a user clicks
"Create a new review", I worry that we're exacerbating the problem we've
always had of users not knowing how to create comments. This change adds
a new carousel of tips to the bottom of the review dialog, explaining
how to use comments and reviews.

Testing Done:
Opened the review dialog for both empty and non-empty reviews. Saw that
when the review was empty, I saw the tips. The tips carousel starts in
auto-cycle mode, and I was able to cycle forward and backward through
the list.

Reviewed at https://reviews.reviewboard.org/r/13173/